Move Order

+++ server/src/main/java/edu/duke/ece651/team13/server/dto/OrderDTO.java
@@ -14,5 +14,5 @@ public class OrderDTO {
     private Long sourceTerritoryId;
     private Long destinationTerritoryId;
     //TODO Update to indicate type of units and add value
-    private Long units;
+    private int units;
 }
