+++ server/src/main/java/edu/duke/ece651/team13/server/entity/OrderEntity.java
@@ -44,5 +44,5 @@ public class OrderEntity {
     private TerritoryEntity destination;
 
     @Column(name = "UNIT_NUM")
-    private long unitNum;
+    private int unitNum;
 }
